+++ apps/extension/src/utils/sentinel.ts
